added interruption parsing to TRFRunner

+++ str_analysis/utils/trf_runner.py
@@ -2,10 +2,12 @@
 Utils for using the TandemRepeatsFinder [Benson 1999] to find repeats in nucleotide sequences.
 """
 
+import collections
 import os
 import re
 import subprocess
 import tempfile
+import traceback
 import unittest
 
 from str_analysis.utils.dat_utils import parse_dat_file
@@ -15,28 +17,43 @@ class TRFRunner:
     """Class for running TandemRepeatFinder on a nucleotide sequence."""
 
     def __init__(self,
-                 trf_command_path,
+                 trf_executable_path,
                  match_score = 2,
                  mismatch_penalty = 7,
                  indel_penalty = 7,
                  pm = 80,
                  pi = 10,
-                 minscore = 24):
+                 minscore = 24, 
+                 output_filename_prefix=None):
+        """
+        Args:
+            trf_executable_path (str): path to the TRF executable
+            match_score (int): see Tandem Repeats Finder docs
+            mismatch_penalty (int): see Tandem Repeats Finder docs
+            indel_penalty (int): see Tandem Repeats Finder docs
+            pm (int): see Tandem Repeats Finder docs
+            pi (int): see Tandem Repeats Finder docs
+            minscore (int): see Tandem Repeats Finder docs
+            output_filename_prefix (str): if specified, the TRF output files will have this prefix. If not specified, intermediate files will be deleted.
+        """
 
-        self.trf_command_path = os.path.expanduser(trf_command_path)
+        self.trf_executable_path = os.path.expanduser(trf_executable_path)
         self.match_score = match_score
         self.mismatch_penalty = mismatch_penalty
         self.indel_penalty = indel_penalty
         self.pm = pm
         self.pi = pi
         self.minscore = minscore
-
+        self.output_filename_prefix = output_filename_prefix
 
     def run_TRF(self, nucleotide_sequence, chromosome_name="chrN", min_motif_size=None, max_motif_size=None):
         """Run TRF on the given nucleotide sequence and return a generator of DatRecords"""
-        temp_fasta_path = _write_sequences_to_temp_fasta({chromosome_name : nucleotide_sequence})
+        temp_fasta_path = _write_sequences_to_temp_fasta(
+            {chromosome_name : nucleotide_sequence},
+            output_filename_prefix=self.output_filename_prefix)
+
         for dat_record in _run_tandem_repeats_finder(
-                self.trf_command_path,
+                self.trf_executable_path,
                 temp_fasta_path,
                 match_score = self.match_score,
                 mismatch_penalty = self.mismatch_penalty,
@@ -47,29 +64,38 @@ class TRFRunner:
                 output_format="dat",
                 min_motif_size=min_motif_size,
                 max_motif_size=max_motif_size,
+                keep_intermediate_files=self.output_filename_prefix is None,
         ):
             yield dat_record
 
-    def run_TRF_and_get_motif_composition(self, nucleotide_sequence, chromosome_name="chrN", min_motif_size=None, max_motif_size=None):
-        """Run TRF on the given nucleotide sequence and return a generator of DatRecords with motif composition"""
-        temp_fasta_path = _write_sequences_to_temp_fasta({chromosome_name : nucleotide_sequence})
+    def run_TRF_and_parse_motif_composition(self, nucleotide_sequence, chromosome_name="chrN", min_motif_size=None, max_motif_size=None):
+        """Run TRF on the given nucleotide sequence and return a generator of records with motif composition"""
+
+        # Write the sequence to a temporary fasta file
+        temp_fasta_path = _write_sequences_to_temp_fasta(
+            {chromosome_name : nucleotide_sequence}, 
+            output_filename_prefix=self.output_filename_prefix)
+
+        # Run TRF and parse the motif composition
         for record in _run_tandem_repeats_finder(
-                self.trf_command_path,
-                temp_fasta_path,
-                match_score = self.match_score,
-                mismatch_penalty = self.mismatch_penalty,
-                indel_penalty = self.indel_penalty,
-                pm = self.pm,
-                pi = self.pi,
-                minscore = self.minscore,
-                output_format="html",
-                min_motif_size=min_motif_size,
-                max_motif_size=max_motif_size,
+            self.trf_executable_path,
+            temp_fasta_path,
+            match_score = self.match_score,
+            mismatch_penalty = self.mismatch_penalty,
+            indel_penalty = self.indel_penalty,
+            pm = self.pm,
+            pi = self.pi,
+            minscore = self.minscore,
+            output_format="html",
+            min_motif_size=min_motif_size,
+            max_motif_size=max_motif_size,
+            output_filename_prefix=self.output_filename_prefix,
+            keep_intermediate_files=self.output_filename_prefix is None,
         ):
             yield record
 
 
-def _write_sequences_to_temp_fasta(sequences):
+def _write_sequences_to_temp_fasta(sequences, output_filename_prefix=None):
     """Create a temporary .fasta file with the given nucleotide sequences.
 
     Args:
@@ -78,12 +104,18 @@ def _write_sequences_to_temp_fasta(sequences):
     Return:
         str: temp file path
     """
-    temp_fasta_file = tempfile.NamedTemporaryFile("w+", suffix=".fasta", delete=False)
+    if output_filename_prefix is None:
+        temp_fasta_file = tempfile.NamedTemporaryFile("w+", suffix=".fasta", delete=False)
+    else:
+        temp_fasta_file = open(f"{output_filename_prefix}.fasta", "wt")
+
     for chromosome_name in sequences:
         temp_fasta_file.write(f">{chromosome_name}\n")
         temp_fasta_file.write(f"{sequences[chromosome_name]}\n")
+    
     temp_fasta_file.close()
 
+    print(f"Wrote {temp_fasta_file.name}")
     return temp_fasta_file.name
 
 
@@ -100,6 +132,7 @@ def _run_tandem_repeats_finder(
         output_filename_prefix=None,
         min_motif_size=None,
         max_motif_size=None,
+        keep_intermediate_files=False,
 ):
     """Runs Tandem Repeats Finder on the given input fasta, parses results and returns a generator of DatRecords
 
@@ -141,12 +174,15 @@ def _run_tandem_repeats_finder(
 
             yield dat_record
 
-        os.remove(output_dat_path)
+        if not keep_intermediate_files:
+            os.remove(output_dat_path)
 
     elif output_format == "html":
 
-        subprocess.run(command, shell=True, stdout=subprocess.DEVNULL, stderr=subprocess.DEVNULL, check=False)
+        subprocess.run(command, shell=True, stdout=subprocess.DEVNULL, check=False)
 
+        print(command)
+        os.system(f"ls -l {os.path.dirname(input_fasta_path)}")
         # read html output
         trf_output_filename_prefix = f"{os.path.basename(input_fasta_path)}.{match_score}.{mismatch_penalty}.{indel_penalty}.{pm}.{pi}.{minscore}.{max_period}"
         i = 1
@@ -156,14 +192,20 @@ def _run_tandem_repeats_finder(
             for record in _parse_trf_html_output(html_file_path, min_motif_size=min_motif_size, max_motif_size=max_motif_size):
                 yield record
 
+            if not keep_intermediate_files:
+                os.remove(html_file_path)
+
             i += 1
             html_file_path =  f"{trf_output_filename_prefix}.{i}.txt.html"
+
+        if not keep_intermediate_files and os.path.isfile(f"{trf_output_filename_prefix}.txt.html"):
+            os.remove(f"{trf_output_filename_prefix}.txt.html")
 
     else:
         raise ValueError(f"Invalid output_format: '{output_format}'. It must be either 'dat' or 'html'.")
 
 
-def _parse_trf_html_output(html_file_path, min_motif_size=None, max_motif_size=None):
+def _parse_trf_html_output(html_file_path, min_motif_size=None, max_motif_size=None, debug=True):
     """Parse the HTML output of TRF and return a list of motifs"""
 
     with open(html_file_path, "rt") as f:
@@ -172,7 +214,6 @@ def _parse_trf_html_output(html_file_path, min_motif_size=None, max_motif_size=N
     results = []
     for i, html_section in enumerate(html_content.split("<A NAME=")[1:]):
         #print(f"Parsing html section #{i+1}")
-
 
         # find the line that starts with "Indices:"
         html_section = html_section.split("Indices: ")[1]
@@ -185,11 +226,9 @@ def _parse_trf_html_output(html_file_path, min_motif_size=None, max_motif_size=N
         if not indices_score_match:
             raise ValueError(f"Could not parse indices and score from line: {lines[0]}")
 
-        start_index = int(indices_score_match.group(1))
-        end_index = int(indices_score_match.group(2))
+        seq_start_index_0based = int(indices_score_match.group(1)) - 1
+        seq_end_index = int(indices_score_match.group(2))
         score = int(indices_score_match.group(3))
-        indices = [start_index - 1, end_index]  # Adjust start index to be 0-based
-
 
         # Parse period, copynumber, and consensus size using regex - example: "Period size: 30  Copynumber: 191.0  Consensus size: 30"
         period_copynumber_consensus_size_match = re.search(r"Period size: (\d+)\s+Copynumber: (\d+\.?\d*)\s+Consensus size: (\d+)", lines[1])
@@ -213,8 +252,12 @@ def _parse_trf_html_output(html_file_path, min_motif_size=None, max_motif_size=N
         lines = lines[3:]
 
         consensus_motif = None
-        motifs = []
-        motifs_and_start_indices = []
+        #motifs = []
+        #motifs_and_start_indices = []
+        motif_positions_with_interruptions = {}  # histogram of motif positions with interruptions
+        total_mismatches = 0
+        total_indels = 0
+
         block_counter = 0
         while i < len(lines):
             start_block_i = i
@@ -228,31 +271,65 @@ def _parse_trf_html_output(html_file_path, min_motif_size=None, max_motif_size=N
                 continue
 
             try:
-                consensus_motif_line = lines_in_block[-1].strip()
-                consensus_motif_match = re.search(r"(\d+)\s+([ACGTN-]+)", consensus_motif_line)
+                consensus_motif_line = lines_in_block[-1]
+                consensus_motif_match = re.search(r"(\d+)\s+([-ACGTN\s]+)", consensus_motif_line)
                 if not consensus_motif_match:
                     raise ValueError(f"Could not parse consensus motif from line: {consensus_motif_line}")
 
-                if consensus_motif is not None and consensus_motif != consensus_motif_match.group(2).replace("-", ""):
+                consensus_motif_sequence_string = consensus_motif_match.group(2)
+                current_consensus_motif = consensus_motif_sequence_string.split(" ")[0]
+                current_consensus_motif_without_deletions = current_consensus_motif.replace("-", "")
+                if consensus_motif is None:
+                    consensus_motif = current_consensus_motif_without_deletions
+                elif consensus_motif != current_consensus_motif_without_deletions:
                     if i + 2 < len(lines):
-                        print(f"WARNING: {html_file_path} consensus motif mismatch: {consensus_motif_match.group(2)} != {consensus_motif} at line: {consensus_motif_line}")
+                        print(f"WARNING: {html_file_path} consensus motif mismatch: {current_consensus_motif_without_deletions} != {consensus_motif} at line: {consensus_motif_line}")
                     continue
 
-                consensus_motif = consensus_motif_match.group(2).replace("-", "")
+                repeat_sequence_line = lines_in_block[-2]
+                repeat_sequence_match = re.search(r"(\d+)\s+([-ACGTN\s]+)", repeat_sequence_line)
+                if not repeat_sequence_match:
+                    raise ValueError(f"Could not parse motif from line: {repeat_sequence_line}")
 
-                motif_line = lines_in_block[-2].strip()
-                #use regex to parse the motif line: 121 GACCCTGACCTTACTAGTTTACAACCACAC
-                motif_match = re.search(r"(\d+)\s+([-ACGTN\s]+)", motif_line)
-                if not motif_match:
-                    raise ValueError(f"Could not parse motif from line: {motif_line}")
-                motif_start_index = int(motif_match.group(1)) - 1
-                motif_sequence = motif_match.group(2).replace("-", "")
-                for motif in motif_sequence.split(" "):
-                    motifs_and_start_indices.append((motif_start_index, motif))
-                    motifs.append(motif)
-                    motif_start_index += len(motif)
+                repeat_sequence_string = repeat_sequence_match.group(2)
+                total_indels += repeat_sequence_string.count("-") + consensus_motif_sequence_string.count("-")
+
+                #repeat_sequence_start_index_0based = int(repeat_sequence_match.group(1)) - 1
+
+                interruptions_string = None
+                if len(lines_in_block) > 2:
+                    interruptions_line = lines_in_block[-3]
+
+                    interruptions_line_offset, _ = repeat_sequence_match.span(2)  # offset of the repeat sequence in the interruptions line
+                    interruptions_string = interruptions_line[interruptions_line_offset:]
+
+                    total_mismatches += interruptions_string.count("*")
+                    if debug:
+                        print(interruptions_string)
+                        print(repeat_sequence_string)
+                        print(consensus_motif_sequence_string)
+
+                interruptions_string_offset = 0
+                for motif in repeat_sequence_string.split(" "):
+                    #motifs.append(motif)
+                    #motifs_and_start_indices.append((repeat_sequence_start_index_0based, motif))
+                    #repeat_sequence_start_index_0based += len(motif)
+    
+                    if interruptions_string is not None:
+                        motif_interruptions = interruptions_string[interruptions_string_offset:interruptions_string_offset + len(motif)]
+                        if debug:
+                            print(f"motif_interruptions: {motif_interruptions}")
+                            print(f"motif              : {motif}")
+                        if "*" in motif_interruptions:
+                            for i, c in enumerate(motif_interruptions):
+                                if c == "*":
+                                    motif_positions_with_interruptions[i + 1] = motif_positions_with_interruptions.get(i + 1, 0) + 1
+
+                    interruptions_string_offset += len(motif) + 1  # add +1 to account for the space between motifs
+
             except Exception as e:
                 print(f"Error parsing {html_file_path} at lines {start_block_i} to {i}: {block_text}\n{e}")
+                traceback.print_exc()
                 continue
 
             block_counter += 1
@@ -263,11 +340,13 @@ def _parse_trf_html_output(html_file_path, min_motif_size=None, max_motif_size=N
             #"motif_size": period,
             #"consensus_motif_size": consensus_size,
             "num_repeats": copynumber,
-            "repeat_sequence_length": indices[1] - indices[0],
-            "motifs": motifs,
-            "motifs_and_start_indices": motifs_and_start_indices,
+            "repeat_sequence_length": seq_end_index - seq_start_index_0based,
+            #"motifs": motifs,
+            #"motifs_and_start_indices": motifs_and_start_indices,
+            "total_mismatches": total_mismatches,
+            "total_indels": total_indels,
             "score": score,
-            #"motif_positions_with_interruptions": motif_positions_with_interruptions,
+            "motif_positions_with_interruptions": motif_positions_with_interruptions,
         })
 
     return results
@@ -285,10 +364,11 @@ def _parse_trf_html_output(html_file_path, min_motif_size=None, max_motif_size=N
 
 
 if __name__ == "__main__":
-    seq = "CAG"*2 + "CTG" + "CAG"*10 + "CTG"*2 + "CAG"*100
+    #seq = "CAG"*2 + "CTG" + "CAG"*10 + "CTG"*2 + "CAG"*100
+    seq = "CAG"*5 + "TAG" + "CTG" + "CTG" + "CTA" + "CAG"*5
 
     r = TRFRunner("~/bin/trf409.macosx")
-    records = list(r.run_TRF_and_get_motif_composition(
+    records = list(r.run_TRF_and_parse_motif_composition(
         nucleotide_sequence=seq,
         chromosome_name="chrN",
     ))
